fix: исправил замечания

+++ ewm-stats-service/service/src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 create table if not exists endpoint_hits
 (
-    id integer generated always as identity primary key,
+    id bigint generated always as identity primary key,
     app varchar not null,
     uri varchar not null,
     ip varchar not null,
